Restrict Result failure side to an Error
Which extends Throwable, allowing us to provide convenient unwrapping
which can throw the failure value if that's how the user would prefer to
consume it.

We only ever placed a non-Error value in the failure slot when it was a
List<Error>, and we only every then used Errors.compose to flatten
those, so that operation has been flattened to always produce the
composed error without the intermediate List<Error> step.